+++ packages/components/pages/f-account-info/test-utils/setup.js
@@ -1,0 +1,3 @@
+export const baseUrl = 'https://smartGatewayBaseUrl.com';
+
+export const token = 'some-auth-token';