+++ packages/components/organisms/f-header/test-utils/helpers/objects.js
@@ -1,0 +1,15 @@
+export const userInfo = {
+    friendlyName: 'John',
+    isAuthenticated: true,
+    email: 'je@test.com',
+    userData: {
+        signupDate: '2018-05-21T22:54:49.4630000Z',
+        email: 'test',
+        'a-UserId': 'test',
+        consumerStatus: 'test'
+    }
+};
+
+export default {
+    userInfo
+};