+++ packages/components/pages/f-account-info/src/constants.js
@@ -1,0 +1,5 @@
+export const EVENT_SPINNER_STOP_LOADING = 'stop-spinner';
+
+export const CONSUMER_DETAILS_URL = 'consumer';
+
+export const CONVERSATION_ID_NAME = 'x-je-conversation';