+++ packages/components/pages/f-account-info/CHANGELOG.md
@@ -4,6 +4,26 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.5.0
+------------------------------
+*November 29, 2021*
+
+### Added
+- Api provider for getting consumer details data.
+
+
+v0.4.0
+------------------------------
+*November 25, 2021*
+
+### Added
+- Send stop f-spinner event for the parent component
+- Expose new isAuthFinished `required` property
+
+### Changed
+- Initialise Page based on isAuthFinished
+
+
 v0.3.0
 ------------------------------
 *November 23, 2021*